moved assert_fdpanel_equal and makeFDPanel out of test_fdpanel.py into pandas.util.testing
